ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
langgraph==0.2.28
langchain-google-genai
google-generativeai
ortools
pydantic==2.7.0
python-dotenv==1.0.0
httpx==0.27.0
motor==3.5.0
asyncpg==0.29.0
scikit-learn==1.5.1
xgboost==2.1.0
joblib==1.4.2
APScheduler==3.10.4
structlog==24.2.0
Faker==25.0.0
numpy==1.26.4
pandas==2.2.2
psycopg2-binary==2.9.9
